Add HPFT008Test
