+++ src/test/java/by/onliner/footerpages/BaseTest.java
@@ -1,0 +1,24 @@
+package by.onliner.footerpages;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+
+public class BaseTest {
+    protected WebDriver driver;
+    private final String BASE_URL = "https://www.onliner.by/";
+    @BeforeEach
+    public void setUp(){
+        ChromeOptions options = new ChromeOptions();
+        options.addArguments("--disable-search-engine-choice-screen");
+        driver = new ChromeDriver(options);
+        driver.manage().window().maximize();
+        driver.get(BASE_URL);
+    }
+    @AfterEach
+    public void tearDown(){
+        driver.quit();
+    }
+}
