+++ src/test/java/by/onliner/footerpages/FT004Test.java
@@ -1,27 +1,12 @@
 package by.onliner.footerpages;
 
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
 
-public class FT004Test {
-    private WebDriver driver;
-    private final String BASE_URL = "https://www.onliner.by/";
 
-    @BeforeEach
-    public void setUp(){
-        ChromeOptions options = new ChromeOptions();
-        options.addArguments("--disable-search-engine-choice-screen");
-        driver = new ChromeDriver(options);
-        driver.manage().window().maximize();
-        driver.get(BASE_URL);
-    }
+public class FT004Test extends BaseTest{
 
     @Test
     public void testWhenUserClickManifestThenManifestOpened(){
@@ -43,8 +28,4 @@ public class FT004Test {
         Assertions.assertEquals(expectedTitleText, actualTitleText);
     }
 
-    @AfterEach
-    public void tearDown(){
-        driver.quit();
-    }
 }
